Add LoadGenesis function
LoadGenesis will calculate the genesis id from the keys in the configuration
file and check it against the id in the configuration.  It will populate the
merkle tree with the genesis claims if it's empty or check that the claims
exist in the merkle tree otherwise.  It returns the claim proofs of the genesis
claims.