Update mt BenchmarkAddEntry to isolate AddEntry
Update `BenchmarkAddEntry` to: First, populate the merkletree, and then
perform benchmarks adding multiple times the same Entry in copies of the
merkletree (instead of adding different Entry in a different tree (as it
was growing) each time). This is done in order to have a more isolated
benchmark of AddEntry.