Add storage.Delete, implement storage with badger
Update storage interface to remove unused methods (Storage.List, Storage.Info,
Tx.Add) and add Tx.Delete method.  Since LevelDbStorage doesn't implement
Tx.Delete, the old interface has been renamed to StorageLegacy and TxLegacy so
that LevelDbStorage is not yet removed.  We were not using the
leveldb.Transaction for Tx, so adding Tx.Delete was not trivial.  While
leveldb.Transaction supports Delete, it doesn't allow concurrent transactions,
which we are using.

Implement the new storage interface with Badger (which supports concurrent
transactions and Delete).

Redefine the MemoryStorage with the Badger implementation (using the InMemory
flag so that nothing is stored on disk).

Resolve #297