Fix #348 approach A
Adds the error managing to the Poseidon Hash calls, instead of giving a panic when the error is reached. Updates the rest of packages on the new usage of the functions where the hash is involved.

The error when calling Poseidon Hash should never be reached, as is only triggered when the inputs of the hash are outside the Finite Field. But, as this is a core package used by other packages, the panic should not exist or at least should not be reachable.

We evaluated also an 'approach B' ( https://github.com/iden3/go-iden3-core/tree/fix/348b ), where the panic still is there but should never be reachable, as the solution is around creating a secure API wrapping ElemBytes in a way that is impossible to create new ones outside the Finite Field (returning an error in the case that is outside), in a way that the panic of the Poseidon Hash call should not be reachable. This approach adds extra complexity in the outside package usage of the ElemBytes, and opens possible holes in the future in the case that something is changed and the panic is reachable again. That's why for the moment we implemented the 'approach A'.