Add ClaimKeyBabyJub, put AuthorizeKSign as SubType
The ClaimKeyBabyJub will be used for all the claims about a Public
BabyJubJub key, while the in the SubType will be specified the usage of
the key, if is a generic key, or if for example is for authorizing the
Public BabyJubJub key to act as the official KSign (KOperational) of the
identity.

In a future work, we can specify the 'ClaimSubType' as a property in the
metadata, so we save space in the elements of the Index of the claim.