Add ClaimKeyBabyJub
Add ClaimKeyBabyJub and put ClaimAuthorizeKSignBabyJub as another type
that is almost the same with the only difference of the ClaimType.

The ClaimKeyBabyJub will be used for all the claims about a Public
BabyJubJub key, while the ClaimAuthorizeKSignBabyJub is specifically to
authorize the Public BabyJubJub key to act as the official KSign (KOperational) of the
identity.