ClaimKeyBabyJub use Ax instead of Sign
ClaimKeyBabyJub use Ax instead of Sign, as we were using the Sign
because we were using only less than 2 elements for the Curve Point, but
now we have much more space.
This change is motivated due compatibility with the circom circuits
implementation, as with this change we save 2 private inputs and some
checks in the circuits that verify the ownership of a key (PbkSign,
PbkAy) for building the ClaimKeyBabyJub, as instead of comparing PbkAy
with the derivated Ay from the PrivateKey, we can directly derive PbkAx
& PbkAy from the PrivateKey.

Also applied the change to ClaimAuthorizeKSignBabyJub.